+++ portal-ui/src/screens/Console/Buckets/Buckets.tsx
@@ -15,13 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import React from "react";
-import {
-  createStyles,
-  makeStyles,
-  Theme,
-  useTheme,
-  withStyles
-} from "@material-ui/core/styles";
+import { createStyles, Theme, withStyles } from "@material-ui/core/styles";
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
@@ -42,12 +36,9 @@ import {
 } from "@material-ui/core";
 import Typography from "@material-ui/core/Typography";
 import DeleteIcon from "@material-ui/icons/Delete";
-import { KeyboardArrowLeft, KeyboardArrowRight } from "@material-ui/icons";
-import { TablePaginationActionsProps } from "@material-ui/core/TablePagination/TablePaginationActions";
-import FirstPageIcon from "@material-ui/icons/FirstPage";
-import LastPageIcon from "@material-ui/icons/LastPage";
 import AddBucket from "./AddBucket";
 import DeleteBucket from "./DeleteBucket";
+import { MinTablePaginationActions } from "../../../common/MinTablePaginationActions";
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -72,86 +63,6 @@ const styles = (theme: Theme) =>
       paddingRight: theme.spacing(0)
     }
   });
-
-const useStyles1 = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      flexShrink: 0,
-      marginLeft: theme.spacing(2.5)
-    }
-  })
-);
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-  const classes = useStyles1();
-  const theme = useTheme();
-  const { count, page, rowsPerPage, onChangePage } = props;
-
-  const handleFirstPageButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, 0);
-  };
-
-  const handleBackButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, page - 1);
-  };
-
-  const handleNextButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <div className={classes.root}>
-      <IconButton
-        onClick={handleFirstPageButtonClick}
-        disabled={page === 0}
-        aria-label="first page"
-      >
-        {theme.direction === "rtl" ? <LastPageIcon /> : <FirstPageIcon />}
-      </IconButton>
-      <IconButton
-        onClick={handleBackButtonClick}
-        disabled={page === 0}
-        aria-label="previous page"
-      >
-        {theme.direction === "rtl" ? (
-          <KeyboardArrowRight />
-        ) : (
-          <KeyboardArrowLeft />
-        )}
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="next page"
-      >
-        {theme.direction === "rtl" ? (
-          <KeyboardArrowLeft />
-        ) : (
-          <KeyboardArrowRight />
-        )}
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="last page"
-      >
-        {theme.direction === "rtl" ? <FirstPageIcon /> : <LastPageIcon />}
-      </IconButton>
-    </div>
-  );
-}
 
 interface IBucketsProps {
   classes: any;
@@ -267,16 +178,16 @@ class Buckets extends React.Component<IBucketsProps, IBucketsState> {
       this.setState({ deleteOpen: true, selectedBucket: bucket });
     };
 
-      const bytesToSize = (bytesStr:string) => {
-          const bytes = parseInt(bytesStr,10)
-          if (isNaN(bytes)) {
-              return '0 Byte';
-          }
-          const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
-          if (bytes === 0) return '0 Byte';
-          const i = Math.floor(Math.log(bytes) / Math.log(1024));
-          return Math.round(bytes / Math.pow(1024, i)) + ' ' + sizes[i];
+    const bytesToSize = (bytesStr: string) => {
+      const bytes = parseInt(bytesStr, 10);
+      if (isNaN(bytes)) {
+        return "0 Byte";
       }
+      const sizes = ["Bytes", "KB", "MB", "GB", "TB"];
+      if (bytes === 0) return "0 Byte";
+      const i = Math.floor(Math.log(bytes) / Math.log(1024));
+      return Math.round(bytes / Math.pow(1024, i)) + " " + sizes[i];
+    };
 
     return (
       <React.Fragment>
@@ -352,11 +263,7 @@ class Buckets extends React.Component<IBucketsProps, IBucketsState> {
               <TableFooter>
                 <TableRow>
                   <TablePagination
-                    rowsPerPageOptions={[
-                      5,
-                      10,
-                      25
-                    ]}
+                    rowsPerPageOptions={[5, 10, 25]}
                     colSpan={3}
                     count={totalRecords}
                     rowsPerPage={rowsPerPage}
@@ -367,7 +274,7 @@ class Buckets extends React.Component<IBucketsProps, IBucketsState> {
                     }}
                     onChangePage={handleChangePage}
                     onChangeRowsPerPage={handleChangeRowsPerPage}
-                    ActionsComponent={TablePaginationActions}
+                    ActionsComponent={MinTablePaginationActions}
                   />
                 </TableRow>
               </TableFooter>
